Removed the utilities state control, and instead modified BioXASUtilities to inherit from BioXASBiStateGroup. It now handles reporting a state value as its own value. #1967
